+++ src/main/java/com/vonage/client/conversations/Event.java
@@ -22,7 +22,9 @@ import java.time.Instant;
 import java.util.Objects;
 
 /**
- * Events are actions that occur within a conversation.
+ * Events are actions that occur within a conversation. This is the base class for all events.
+ * The specific event subtype is determined by the {@code type} field.
+ * For custom events, use the {@linkplain #getTypeName()} to determine the specific custom suffix.
  */
 @JsonTypeInfo(
 		use = JsonTypeInfo.Id.NAME,
@@ -49,9 +51,12 @@ import java.util.Objects;
 	@JsonSubTypes.Type(value = AudioSpeakingOffEvent.class, name = "audio:speaking:off"),
 	@JsonSubTypes.Type(value = AudioSpeakingOnEvent.class, name = "audio:speaking:on"),
 	@JsonSubTypes.Type(value = ConversationUpdatedEvent.class, name = "conversation:updated"),
-	@JsonSubTypes.Type(value = CustomEvent.class, name = "custom"),
+	@JsonSubTypes.Type(value = CustomEvent.class, name = "custom:"),
 	@JsonSubTypes.Type(value = EphemeralEvent.class, name = "ephemeral"),
 	@JsonSubTypes.Type(value = EventDeleteEvent.class, name = "event:delete"),
+	@JsonSubTypes.Type(value = MemberInvitedEvent.class, name = "member:invited"),
+	@JsonSubTypes.Type(value = MemberJoinedEvent.class, name = "member:joined"),
+	@JsonSubTypes.Type(value = MemberLeftEvent.class, name = "member:left"),
 	@JsonSubTypes.Type(value = MemberMediaEvent.class, name = "member:media"),
 	@JsonSubTypes.Type(value = MessageDeliveredEvent.class, name = "message:delivered"),
 	@JsonSubTypes.Type(value = MessageRejectedEvent.class, name = "message:rejected"),
@@ -60,17 +65,19 @@ import java.util.Objects;
 	@JsonSubTypes.Type(value = MessageUndeliverableEvent.class, name = "message:undeliverable"),
 	@JsonSubTypes.Type(value = RtcAnswerEvent.class, name = "rtc:answer"),
 	@JsonSubTypes.Type(value = RtcAnsweredEvent.class, name = "rtc:answered"),
+	@JsonSubTypes.Type(value = RtcHangupEvent.class, name = "rtc:hangup"),
 	@JsonSubTypes.Type(value = RtcRingingEvent.class, name = "rtc:ringing"),
 	@JsonSubTypes.Type(value = RtcStatusEvent.class, name = "rtc:status"),
 	@JsonSubTypes.Type(value = RtcTransferEvent.class, name = "rtc:transfer"),
 	@JsonSubTypes.Type(value = SipAmdMachineEvent.class, name = "sip:amd_machine"),
 	@JsonSubTypes.Type(value = SipAnsweredEvent.class, name = "sip:answered"),
+	@JsonSubTypes.Type(value = SipHangupEvent.class, name = "sip:hangup"),
 	@JsonSubTypes.Type(value = SipMachineEvent.class, name = "sip:machine"),
 	@JsonSubTypes.Type(value = SipRingingEvent.class, name = "sip:ringing")
 })
 public abstract class Event extends JsonableBaseObject {
 	@JsonIgnore String conversationId;
-	@JsonProperty("type") EventType type;
+	@JsonProperty("type") String type;
 	@JsonProperty("id") Integer id;
 	@JsonProperty("from") String from;
 	@JsonProperty("timestamp") Instant timestamp;
@@ -103,7 +110,19 @@ public abstract class Event extends JsonableBaseObject {
 	 * 
 	 * @return The event type as an enum.
 	 */
+	@JsonIgnore
 	public EventType getType() {
+		return EventType.fromString(type);
+	}
+
+	/**
+	 * Gets the event type as a string. This is mainly useful for distinguishing {@link EventType#CUSTOM} events.
+	 *
+	 * @return The event type as a string, including the suffix if it's a custom event.
+	 * @since 8.20.0
+	 */
+	@JsonProperty("type")
+	public String getTypeName() {
 		return type;
 	}
 
@@ -153,8 +172,19 @@ public abstract class Event extends JsonableBaseObject {
 	 */
 	@SuppressWarnings("unchecked")
 	public abstract static class Builder<E extends Event, B extends Builder<? extends E, ? extends B>> {
-		private final EventType type;
+		private final String type;
 		private String from;
+
+		/**
+		 * Construct a new builder for a given event type.
+		 *
+		 * @param type The event type as a string.
+		 * @since 8.20.0
+		 */
+		protected Builder(String type) {
+			// Validate the event type.
+			EventType.fromString(this.type = Objects.requireNonNull(type, "Event type is required."));
+		}
 
 		/**
 		 * Construct a new builder for a given event type.
@@ -162,7 +192,7 @@ public abstract class Event extends JsonableBaseObject {
 		 * @param type The event type as an enum.
 		 */
 		protected Builder(EventType type) {
-			this.type = type;
+			this(type.toString());
 		}
 
 		/**